Traslada funcion de actualizar datos de organizacion a un job
