+++ jobs/actualizarOrganizacionesJob.ts
@@ -1,0 +1,10 @@
+import { actualizarOfertasPrestacionales } from './actualizarOrganizaciones';
+import { actualizarContacto } from './actualizarOrganizaciones';
+
+async function run(done) {
+    await actualizarOfertasPrestacionales();
+    await actualizarContacto();
+    done();
+}
+
+export = run;
